+++ Src/DependencyCollector/Net45.Tests/DependencyTrackingTelemetryModuleHttpTest.cs
@@ -87,6 +87,13 @@
 
         [TestMethod]
         [Timeout(5000)]
+        public void TestBasicDependencyCollectionDiagnosticSourceLegacyHeaders()
+        {
+            this.TestCollectionSuccessfulResponse(true, LocalhostUrlDiagSource, 200, true);
+        }
+
+        [TestMethod]
+        [Timeout(5000)]
         public async Task TestZeroContentResponseDiagnosticSource()
         {
             await this.TestCollectionHttpClientSucessfulResponse(LocalhostUrlDiagSource, 200, 0);
@@ -147,11 +154,9 @@
                 // HttpDesktopDiagnosticListener cannot collect dependencies if HttpWebResponse was not closed/disposed
                 Assert.IsFalse(this.sentTelemetry.Any());
                 var requestId = request.Headers[RequestResponseHeaders.RequestIdHeader];
-                var rootId = request.Headers[RequestResponseHeaders.StandardRootIdHeader];
                 Assert.IsNotNull(requestId);
-                Assert.IsNotNull(rootId);
-                Assert.AreEqual(requestId, request.Headers[RequestResponseHeaders.StandardParentIdHeader]);
-                Assert.IsTrue(requestId.StartsWith('|' + rootId + '.'));
+                Assert.IsNull(request.Headers[RequestResponseHeaders.StandardRootIdHeader]);
+                Assert.IsNull(request.Headers[RequestResponseHeaders.StandardParentIdHeader]);
             }
         }
 
@@ -336,9 +341,9 @@
             }
         }
 
-        private void TestCollectionSuccessfulResponse(bool enableDiagnosticSource, string url, int statusCode)
+        private void TestCollectionSuccessfulResponse(bool enableDiagnosticSource, string url, int statusCode, bool injectLegacyHeaders = false)
         {
-            using (this.CreateDependencyTrackingModule(enableDiagnosticSource))
+            using (this.CreateDependencyTrackingModule(enableDiagnosticSource, injectLegacyHeaders))
             {
                 HttpWebRequest request = WebRequest.CreateHttp(url);
 
@@ -370,11 +375,11 @@
                     }
                 }
 
-                this.ValidateTelemetry(enableDiagnosticSource, this.sentTelemetry.Single(), new Uri(url), request, statusCode >= 200 && statusCode < 300, statusCode.ToString(CultureInfo.InvariantCulture));
+                this.ValidateTelemetry(enableDiagnosticSource, this.sentTelemetry.Single(), new Uri(url), request, statusCode >= 200 && statusCode < 300, statusCode.ToString(CultureInfo.InvariantCulture), injectLegacyHeaders);
             }
         }
 
-        private async Task TestCollectionHttpClientSucessfulResponse(string url, int statusCode, int contentLength)
+        private async Task TestCollectionHttpClientSucessfulResponse(string url, int statusCode, int contentLength, bool injectLegacyHeaders = false)
         {
             using (this.CreateDependencyTrackingModule(true))
             {
@@ -458,7 +463,7 @@
             }
         }
 
-        private void ValidateTelemetry(bool diagnosticSource, DependencyTelemetry item, Uri url, WebRequest request, bool success, string resultCode)
+        private void ValidateTelemetry(bool diagnosticSource, DependencyTelemetry item, Uri url, WebRequest request, bool success, string resultCode, bool expectLegacyHeaders = false)
         {
             Assert.AreEqual(url, item.Data);
 
@@ -489,7 +494,7 @@
 
             if (diagnosticSource)
             {
-                this.ValidateTelemetryForDiagnosticSource(item, url, request);
+                this.ValidateTelemetryForDiagnosticSource(item, url, request, expectLegacyHeaders);
             }
             else
             {
@@ -497,7 +502,7 @@
             }
         }
 
-        private void ValidateTelemetryForDiagnosticSource(DependencyTelemetry item, Uri url, WebRequest request)
+        private void ValidateTelemetryForDiagnosticSource(DependencyTelemetry item, Uri url, WebRequest request, bool expectLegacyHeaders)
         {
             var expectedMethod = request != null ? request.Method : "GET";
             Assert.AreEqual(expectedMethod + " " + url.AbsolutePath, item.Name);
@@ -511,8 +516,17 @@
             if (request != null)
             {
                 Assert.AreEqual(requestId, request.Headers[RequestResponseHeaders.RequestIdHeader]);
-                Assert.AreEqual(requestId, request.Headers[RequestResponseHeaders.StandardParentIdHeader]);
-                Assert.AreEqual(item.Context.Operation.Id, request.Headers[RequestResponseHeaders.StandardRootIdHeader]);
+
+                if (expectLegacyHeaders)
+                {
+                    Assert.AreEqual(requestId, request.Headers[RequestResponseHeaders.StandardParentIdHeader]);
+                    Assert.AreEqual(item.Context.Operation.Id, request.Headers[RequestResponseHeaders.StandardRootIdHeader]);
+                }
+                else
+                {
+                    Assert.IsNull(request.Headers[RequestResponseHeaders.StandardParentIdHeader]);
+                    Assert.IsNull(request.Headers[RequestResponseHeaders.StandardRootIdHeader]);
+                }
 
                 if (Activity.Current != null)
                 {
@@ -551,7 +565,7 @@
             }
         }
 
-        private DependencyTrackingTelemetryModule CreateDependencyTrackingModule(bool enableDiagnosticSource)
+        private DependencyTrackingTelemetryModule CreateDependencyTrackingModule(bool enableDiagnosticSource, bool injectLegacyHeaders = false)
         {
             var module = new DependencyTrackingTelemetryModule();
 
@@ -559,6 +573,8 @@
             {
                 module.DisableDiagnosticSourceInstrumentation = true;
             }
+
+            module.EnableLegacyCorrelationHeadersInjection = injectLegacyHeaders;
 
             module.Initialize(this.config);
             Assert.AreEqual(enableDiagnosticSource, DependencyTableStore.IsDesktopHttpDiagnosticSourceActivated);