+++ Src/DependencyCollector/Shared/HttpCoreDiagnosticSourceListener.cs
@@ -54,9 +54,10 @@ namespace Microsoft.ApplicationInsights.DependencyCollector.Implementation
         private readonly ConcurrentDictionary<string, Exception> pendingExceptions =
             new ConcurrentDictionary<string, Exception>();
 
-        private bool isNetCore20HttpClient;
+        private readonly bool isNetCore20HttpClient;
+        private readonly bool injectLegacyHeaders = false;
 
-        public HttpCoreDiagnosticSourceListener(TelemetryConfiguration configuration, bool setComponentCorrelationHttpHeaders, IEnumerable<string> correlationDomainExclusionList)
+        public HttpCoreDiagnosticSourceListener(TelemetryConfiguration configuration, bool setComponentCorrelationHttpHeaders, IEnumerable<string> correlationDomainExclusionList, bool injectLegacyHeaders)
         {
             this.client = new TelemetryClient(configuration);
             this.client.Context.GetInternalContext().SdkVersion = SdkVersionUtils.GetSdkVersion("rdd" + RddSource.DiagnosticSourceCore + ":");
@@ -68,6 +69,7 @@ namespace Microsoft.ApplicationInsights.DependencyCollector.Implementation
             this.applicationInsightsUrlFilter = new ApplicationInsightsUrlFilter(configuration);
             this.setComponentCorrelationHttpHeaders = setComponentCorrelationHttpHeaders;
             this.correlationDomainExclusionList = correlationDomainExclusionList ?? Enumerable.Empty<string>();
+            this.injectLegacyHeaders = injectLegacyHeaders;
 
             this.subscriber = new HttpCoreDiagnosticSourceSubscriber(this, this.applicationInsightsUrlFilter, this.isNetCore20HttpClient);
         }
@@ -435,42 +437,48 @@ namespace Microsoft.ApplicationInsights.DependencyCollector.Implementation
                         AppMapCorrelationEventSource.Log.UnknownError(ExceptionUtilities.GetExceptionDetailString(e));
                     }
 
-                    // Add the root ID
-                    string rootId = currentActivity.RootId;
-                    if (!string.IsNullOrEmpty(rootId) &&
-                        !requestHeaders.Contains(RequestResponseHeaders.StandardRootIdHeader))
+                    if (isLegacyEvent)
                     {
-                        requestHeaders.Add(RequestResponseHeaders.StandardRootIdHeader, rootId);
-                    }
-
-                    // Add the parent ID
-                    string parentId = currentActivity.Id;
-                    if (!string.IsNullOrEmpty(parentId) &&
-                        !requestHeaders.Contains(RequestResponseHeaders.StandardParentIdHeader))
-                    {
-                        requestHeaders.Add(RequestResponseHeaders.StandardParentIdHeader, parentId);
-                        if (isLegacyEvent)
+                        if (!requestHeaders.Contains(RequestResponseHeaders.RequestIdHeader))
                         {
-                            requestHeaders.Add(RequestResponseHeaders.RequestIdHeader, parentId);
+                            requestHeaders.Add(RequestResponseHeaders.RequestIdHeader, currentActivity.Id);
+                        }
+
+                        if (!requestHeaders.Contains(RequestResponseHeaders.CorrelationContextHeader))
+                        {
+                            // we expect baggage to be empty or contain a few items
+                            using (IEnumerator<KeyValuePair<string, string>> e = currentActivity.Baggage.GetEnumerator())
+                            {
+                                if (e.MoveNext())
+                                {
+                                    var baggage = new List<string>();
+                                    do
+                                    {
+                                        KeyValuePair<string, string> item = e.Current;
+                                        baggage.Add(new NameValueHeaderValue(item.Key, item.Value).ToString());
+                                    }
+                                    while (e.MoveNext());
+
+                                    requestHeaders.Add(RequestResponseHeaders.CorrelationContextHeader, baggage);
+                                }
+                            }
                         }
                     }
 
-                    if (isLegacyEvent)
+                    if (this.injectLegacyHeaders)
                     {
-                        // we expect baggage to be empty or contain a few items
-                        using (IEnumerator<KeyValuePair<string, string>> e = currentActivity.Baggage.GetEnumerator())
+                        // Add the root ID
+                        string rootId = currentActivity.RootId;
+                        if (!string.IsNullOrEmpty(rootId) && !requestHeaders.Contains(RequestResponseHeaders.StandardRootIdHeader))
                         {
-                            if (e.MoveNext())
-                            {
-                                var baggage = new List<string>();
-                                do
-                                {
-                                    KeyValuePair<string, string> item = e.Current;
-                                    baggage.Add(new NameValueHeaderValue(item.Key, item.Value).ToString());
-                                }
-                                while (e.MoveNext());
-                                request.Headers.Add(RequestResponseHeaders.CorrelationContextHeader, baggage);
-                            }
+                            requestHeaders.Add(RequestResponseHeaders.StandardRootIdHeader, rootId);
+                        }
+
+                        // Add the parent ID
+                        string parentId = currentActivity.Id;
+                        if (!string.IsNullOrEmpty(parentId) && !requestHeaders.Contains(RequestResponseHeaders.StandardParentIdHeader))
+                        {
+                            requestHeaders.Add(RequestResponseHeaders.StandardParentIdHeader, parentId);
                         }
                     }
                 }