+++ Src/DependencyCollector/Shared/TelemetryExtensionsForDependencyCollector.cs
@@ -1,5 +1,6 @@
 ﻿namespace Microsoft.ApplicationInsights.DependencyCollector
 {
+    using System;
     using System.Data.SqlClient;
     using System.Net;
     using Microsoft.ApplicationInsights.DataContracts;
@@ -19,44 +20,9 @@
         /// <param name="setCookies">Set cookies enables the process of setting the cookies to the web request. By default it is set to false.</param>
         /// <param name="setCorrelationContext">Set request headers to correlate dependency telemetry item with the request telemetry item that will process this http request.</param>
         /// <returns>Dependency telemetry item with an associated dependency telemetry item.</returns>
+        [Obsolete("This method is obsolete. Http dependency tracking, correlation and HTTP header injection are supported out of the box. Use TelemetryClient.StartOperation for manual tracking and correlation", true)]
         public static DependencyTelemetry AssociateTelemetryWithWebRequest(this DependencyTelemetry telemetry, WebRequest webRequest, bool setCookies = false, bool setCorrelationContext = false)
         {
-            if (webRequest == null)
-            {
-                DependencyCollectorEventSource.Log.WebRequestIsNullWarning();
-                return null;
-            }
-
-            if (telemetry == null)
-            {
-                DependencyCollectorEventSource.Log.TelemetryToTrackIsNullWarning();
-                return null;
-            }
-
-            DependencyTelemetry result;
-
-            var telemetryTuple = ClientServerDependencyTracker.GetTupleForWebDependencies(webRequest);
-
-            if (telemetryTuple != null)
-            {
-                result = telemetryTuple.Item1 as DependencyTelemetry;
-            }
-            else
-            {
-                result = telemetry;
-                ClientServerDependencyTracker.AddTupleForWebDependencies(webRequest, telemetry, true);
-            }
-
-            if (setCookies)
-            {
-                WebRequestDependencyTrackingHelpers.SetUserAndSessionContextForWebRequest(result, webRequest);
-            }
-
-            if (setCorrelationContext)
-            {
-                WebRequestDependencyTrackingHelpers.SetCorrelationContextForWebRequest(result, webRequest);
-            }
-
             return telemetry;
         }
 
@@ -67,31 +33,9 @@
         /// <param name="telemetry">Telemetry object that needs to be associated with the web request.</param>
         /// <param name="sqlRequest">SQL request object which is used as a key to store in the tables.</param>
         /// <returns>Dependency telemetry item with an associated dependency telemetry item.</returns>
+        [Obsolete("This method is obsolete. Sql depednency tracking and correaltion is supported out of the box. Use TelemetryClient.StartOperation for manual tracking and correlation", true)]
         public static DependencyTelemetry AssociateTelemetryWithSqlRequest(this DependencyTelemetry telemetry, SqlCommand sqlRequest)
         {
-            if (sqlRequest == null)
-            {
-                DependencyCollectorEventSource.Log.WebRequestIsNullWarning();
-                return null;
-            }
-
-            if (telemetry == null)
-            {
-                DependencyCollectorEventSource.Log.TelemetryToTrackIsNullWarning();
-                return null;
-            }
-
-            var telemetryTuple = ClientServerDependencyTracker.GetTupleForSqlDependencies(sqlRequest);
-
-            if (telemetryTuple != null)
-            {
-                return telemetryTuple.Item1 as DependencyTelemetry;
-            }
-            else
-            {
-                ClientServerDependencyTracker.AddTupleForSqlDependencies(sqlRequest, telemetry, true);
-            }
-
             return telemetry;
         }
     }