+++ Src/Web/Web.Net45.Tests/AspNetDiagnosticTelemetryModuleTest.cs
@@ -177,7 +177,7 @@
         }
 
         [TestMethod]
-        public void RequestTelemetryIsSetWithLegacyHeaders()
+        public void RequestTelemetryIsNotSetWithLegacyHeaders()
         {
             FakeAspNetDiagnosticSource.FakeContext =
                 HttpModuleHelper.GetFakeHttpContext(new Dictionary<string, string>
@@ -187,6 +187,31 @@
                 });
 
             this.module = this.CreateModule();
+
+            var activity = new Activity(FakeAspNetDiagnosticSource.IncomingRequestEventName);
+            this.aspNetDiagnosticsSource.StartActivityWithoutChecks(activity);
+            this.aspNetDiagnosticsSource.StopActivity();
+
+            Assert.AreEqual(1, this.sendItems.Count);
+
+            var requestTelemetry = this.sendItems[0] as RequestTelemetry;
+            Assert.IsNotNull(requestTelemetry);
+            Assert.AreEqual(activity.RootId, requestTelemetry.Context.Operation.Id);
+            Assert.IsNull(requestTelemetry.Context.Operation.ParentId);
+            Assert.AreEqual(activity.Id, requestTelemetry.Id);
+        }
+
+        [TestMethod]
+        public void RequestTelemetryIsSetWithLegacyHeaders()
+        {
+            FakeAspNetDiagnosticSource.FakeContext =
+                HttpModuleHelper.GetFakeHttpContext(new Dictionary<string, string>
+                {
+                    ["x-ms-request-id"] = "guid1",
+                    ["x-ms-request-root-id"] = "guid2"
+                });
+
+            this.module = this.CreateModule("x-ms-request-root-id", "x-ms-request-id");
 
             var activity = new Activity(FakeAspNetDiagnosticSource.IncomingRequestEventName);
             Assert.IsTrue(this.aspNetDiagnosticsSource.IsEnabled(FakeAspNetDiagnosticSource.IncomingRequestEventName, activity));